refactor: remove top-level compat shims, use mesh.session/mesh.audit directly
- Delete strands_robots/mesh_session.py and strands_robots/mesh_audit.py
- Update all test imports to use strands_robots.mesh.session / .audit
- No backward-compat shims needed — mesh/ package is the canonical path
