--- strands_robots/mesh_audit.py
@@ -1,3 +1,0 @@
-"""Backward-compat shim — re-exports from strands_robots.mesh.audit."""
-
-from strands_robots.mesh.audit import *  # noqa: F401,F403
